create `clear-x` modules, to put into repos later
