+++ component.json
@@ -56,8 +56,8 @@
     "component/assert": "*",
     "jkroso/equals": "*",
     "matthewp/keys": "*",
-    "segmentio/analytics.js": "*",
-    "segmentio/analytics.js-integration-tester": "*",
+    "segmentio/analytics.js": "add/property",
+    "segmentio/analytics.js-integration-tester": "new/loaded",
     "segmentio/spy": "*",
     "segmentio/when": "*",
     "visionmedia/mocha": "*",
