+++ lib/twitter-ads/test.js
@@ -41,6 +41,7 @@ describe('Twitter Ads', function(){
     beforeEach(function(done){
       analytics.once('ready', done);
       analytics.initialize();
+      
       analytics.page();
     });
